#10 Minor cleanups.
