+++ src/main/java/io/github/galop_proxy/galop/configuration/HttpHeaderRequestConfiguration.java
@@ -1,9 +1,5 @@
 package io.github.galop_proxy.galop.configuration;
 
-public interface HttpHeaderRequestConfiguration {
-
-    long getReceiveTimeout();
-
-    int getMaxSize();
+public interface HttpHeaderRequestConfiguration extends HttpHeaderTypeConfiguration {
 
 }
