+++ src/main/java/io/github/galop_proxy/galop/configuration/HttpHeaderRequestConfigurationImpl.java
@@ -1,23 +1,10 @@
 package io.github.galop_proxy.galop.configuration;
 
-final class HttpHeaderRequestConfigurationImpl implements HttpHeaderRequestConfiguration {
-
-    private final long receiveTimeout;
-    private final int maxSize;
+final class HttpHeaderRequestConfigurationImpl extends AbstractHttpHeaderTypeConfiguration
+        implements HttpHeaderRequestConfiguration {
 
     HttpHeaderRequestConfigurationImpl(final long receiveTimeout, final int maxSize) {
-        this.receiveTimeout = receiveTimeout;
-        this.maxSize = maxSize;
-    }
-
-    @Override
-    public long getReceiveTimeout() {
-        return receiveTimeout;
-    }
-
-    @Override
-    public int getMaxSize() {
-        return maxSize;
+        super(receiveTimeout, maxSize);
     }
 
 }
