+++ src/main/java/io/github/galop_proxy/galop/configuration/HttpHeaderResponseConfiguration.java
@@ -1,9 +1,5 @@
 package io.github.galop_proxy.galop.configuration;
 
-public interface HttpHeaderResponseConfiguration {
-
-    long getReceiveTimeout();
-
-    int getMaxSize();
+public interface HttpHeaderResponseConfiguration extends HttpHeaderTypeConfiguration {
 
 }
