+++ src/main/java/io/github/galop_proxy/galop/configuration/HttpHeaderResponseConfigurationImpl.java
@@ -1,23 +1,10 @@
 package io.github.galop_proxy.galop.configuration;
 
-final class HttpHeaderResponseConfigurationImpl implements HttpHeaderResponseConfiguration {
-
-    private final long receiveTimeout;
-    private final int maxSize;
+final class HttpHeaderResponseConfigurationImpl extends AbstractHttpHeaderTypeConfiguration
+        implements HttpHeaderResponseConfiguration {
 
     HttpHeaderResponseConfigurationImpl(final long receiveTimeout, final int maxSize) {
-        this.receiveTimeout = receiveTimeout;
-        this.maxSize = maxSize;
-    }
-
-    @Override
-    public long getReceiveTimeout() {
-        return receiveTimeout;
-    }
-
-    @Override
-    public int getMaxSize() {
-        return maxSize;
+        super(receiveTimeout, maxSize);
     }
 
 }
